requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
sqlalchemy==2.0.31
qrcode[pil]==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
